pyproject.toml
[tool.poetry]
name = "task_api_project"
version = "0.1.0"
description = "Task Management API built with FastAPI, SQLAlchemy, Alembic, JWT auth"
authors = ["Begli Geldiyev <begli.geld.08.12.01@gmail.com>"]
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.2"
uvicorn = {extras = ["standard"], version = "^0.23.0"}
SQLAlchemy = "^1.4"
asyncpg = "^0.30.0"
alembic = "^1.11.1"
pydantic = {version = "^1.10.7", extras = ["email"]}
PyJWT = "^2.6.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-dotenv = "^1.0.0"
python-multipart = "^0.0.20"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
httpx = "^0.24.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
